Move help text helpers to separate file
