+++ frontend/src/metabase-lib/v1/expressions/index.ts
@@ -6,3 +6,4 @@ export * from "./diagnostics";
 export * from "./position";
 export * from "./errors";
 export * from "./clause";
+export * from "./help-text";
