+++ frontend/src/metabase/query_builder/components/expressions/FunctionBrowser/utils.ts
@@ -7,8 +7,8 @@ import {
   EXPRESSION_FUNCTIONS,
   type HelpText,
   type MBQLClauseFunctionConfig,
+  getHelpText,
 } from "metabase-lib/v1/expressions";
-import { getHelpText } from "metabase-lib/v1/expressions/helper-text-strings";
 import type Database from "metabase-lib/v1/metadata/Database";
 import type Metadata from "metabase-lib/v1/metadata/Metadata";
